# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Paqtra API
  version: "2.0.0"
  description: >
    Paqtra — trace every flow. Network observability and security for
    Cilium, Hubble, and eBPF on Kubernetes. Real-time flow monitoring,
    anomaly detection, policy management, zero-trust scoring, multi-cluster
    visibility, cost analysis, and automated troubleshooting.
  contact:
    name: Paqtra Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
    description: Local server (health and metrics live at the root)
  - url: /api/v1
    description: Primary API base path
...
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# Paths
# ──────────────────────────────────────────────────────────────
paths:
...
  /api/v1/modules/canary/{id}:
    get:
      operationId: canaryStatus
      tags: [Modules]
      summary: Get canary deployment status
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Canary status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanaryStatus"

  # ── Replay ──────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT token issued by the cluster identity provider. Include in the
        Authorization header as `Bearer <token>`.

  # ── Reusable Parameters ─────────────────────────────────────
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Unique resource identifier
...
  schemas:
...
    CanaryStatus:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        status:
          type: string
          enum: [progressing, succeeded, failed, paused]
        weight:
          type: integer
          description: Percentage of traffic routed to canary
        success_rate:
          type: number
        latency_p99_ms:
          type: number
        iterations:
          type: integer

    # ── Replay ─────────────────────────────────────────────────
...